add links for translated code
Recently, while I'm reading the book, I ported the code to Python and Go to help myself understand the algorithms and thought it perhaps can also help others as well.
So I wonder whether I can put the links to this repo to do a little contribution.

Feel free to close this if it's not suitable. Either way thanks again for the great book.